feat/타입 관리 폴더 추가
