+++ tsconfig.json
@@ -18,6 +18,6 @@
     "jsx": "react-jsx",
     "types": ["react/next", "react-dom/next"]
   },
-  "include": ["src"],
+  "include": ["src", "src/@types"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
